+++ broker/lib/fabrik/DirectorManager.js
@@ -10,7 +10,7 @@ const bosh = require('../../../data-access-layer/bosh');
 const cf = require('../../../data-access-layer/cf');
 const backupStore = require('../../../data-access-layer/iaas').backupStore;
 const utils = require('../../../common/utils');
-const Agent = require('./Agent');
+const Agent = require('../../../data-access-layer/service-agent');
 const BaseManager = require('./BaseManager');
 const DirectorInstance = require('./DirectorInstance');
 const CONST = require('../../../common/constants');
