+++ broker/lib/fabrik/Fabrik.js
@@ -4,7 +4,7 @@ const assert = require('assert');
 const Promise = require('bluebird');
 const config = require('../../../common/config');
 const catalog = require('../../../common/models').catalog;
-const Agent = require('./Agent');
+const Agent = require('../../../data-access-layer/service-agent');
 const VirtualHostAgent = require('./VirtualHostAgent');
 const DirectorManager = require('./DirectorManager');
 const VirtualHostManager = require('./VirtualHostManager');
