+++ broker/lib/fabrik/OobBackupManager.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const Promise = require('bluebird');
 const logger = require('../../../common/logger');
 const backupStore = require('../../../data-access-layer/iaas').backupStoreForOob;
-const Agent = require('./Agent');
+const Agent = require('../../../data-access-layer/service-agent');
 const utils = require('../../../common/utils');
 const errors = require('../../../common/errors');
 const bosh = require('../../../data-access-layer/bosh');
