+++ data-access-layer/service-agent/Agent.js
@@ -4,9 +4,9 @@ const _ = require('lodash');
 const formatUrl = require('url').format;
 const path = require('path');
 const Promise = require('bluebird');
-const utils = require('../../../common/utils');
-const logger = require('../../../common/logger');
-const errors = require('../../../common/errors');
+const utils = require('../../common/utils');
+const logger = require('../../common/logger');
+const errors = require('../../common/errors');
 const HttpClient = utils.HttpClient;
 const compareVersions = utils.compareVersions;
 const FeatureNotSupportedByAnyAgent = errors.FeatureNotSupportedByAnyAgent;
