+++ managers/backup-manager/BackupService.js
@@ -9,7 +9,7 @@ const bosh = require('../../data-access-layer/bosh');
 const backupStore = require('../../data-access-layer/iaas').backupStore;
 const utils = require('../../common/utils');
 const eventmesh = require('../../data-access-layer/eventmesh');
-const Agent = require('../../broker/lib/fabrik/Agent');
+const Agent = require('../../data-access-layer/service-agent');
 const ScheduleManager = require('../../jobs');
 const CONST = require('../../common/constants');
 const BaseDirectorService = require('../BaseDirectorService');
