+++ test/test_broker/fabrik.OOBBackupManager.spec.js
@@ -4,7 +4,7 @@ const Promise = require('bluebird');
 const _ = require('lodash');
 const BoshDirectorClient = require('../../data-access-layer/bosh/BoshDirectorClient');
 const BackupStore = require('../../data-access-layer/iaas/BackupStore');
-const Agent = require('../../broker/lib/fabrik/Agent');
+const Agent = require('../../data-access-layer/service-agent');
 const FabrikBaseController = require('../../api-controllers/FabrikBaseController');
 const CONST = require('../../common/constants');
 const OobBackupManager = require('../../broker/lib/fabrik/OobBackupManager');
